tested send email function with .env var added, fixed issue with .env not passing var till server was running
