+++ controllers/home-routes.js
@@ -1,5 +1,7 @@
 const router = require("express").Router();
 const { Beta, Rating, Food, Mange } = require("../models");
+const sendEmail = require("../utils/emails");
+
 router.get("/", async (req, res) => {
   try {
     const betatasterData = await Beta.findAll({}).catch((err) => {
@@ -9,6 +11,7 @@ router.get("/", async (req, res) => {
     const betatasters = betatasterData.map((betatasters) =>
       betatasters.get({ plain: true })
     );
+    sendEmail();
 
     res.render("review-user", {
       betatasters,
